move custom icons to features/icons
